+++ crates/solana-utils/src/lib.rs
@@ -22,9 +22,11 @@ pub mod compute_budget;
 pub mod transaction_builder;
 
 /// Transaction bundle builder.
+#[cfg(client)]
 pub mod bundle_builder;
 
 /// RPC client extension.
+#[cfg(client)]
 pub(crate) mod client;
 
 /// Utils.
@@ -35,7 +37,8 @@ pub use crate::error::Error;
 /// Result type.
 pub type Result<T> = std::result::Result<T, Error>;
 
-#[cfg(feature = "anchor")]
+#[cfg(anchor)]
 pub use anchor_lang;
+#[cfg(client)]
 pub use solana_client;
 pub use solana_sdk;